:recycle: Refactor minor changes to keep #635 clean
