+++ tests/engines/_test_semantic_segmentation.py
@@ -289,7 +289,7 @@ def test_crash_segmentor(remote_sample: Callable) -> None:
             units="baseline",
         )
 
-    _rm_dir("output")
+    shutil.rmtree("output", ignore_errors=True)
 
     with pytest.raises(ValueError, match=r"Invalid resolution.*"):
         semantic_segmentor.predict(
